Refactor to ColorPicker component
